requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
iopath==0.1.10
jinja2==3.1.5
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
packaging==24.2
pillow==11.1.0
portalocker==3.1.1
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
pyyaml==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
-e file:///home/adnan/vizcom-demo/sam2
setuptools==75.8.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typer==0.15.1
typing-extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
